+++ src/main/kotlin/com/viartemev/thewhiterabbit/channel/Channel.kt
@@ -1,0 +1,9 @@
+package com.viartemev.thewhiterabbit.channel
+
+import com.rabbitmq.client.Channel
+import com.viartemev.thewhiterabbit.consumer.ConfirmConsumer
+
+fun Channel.consumer(queue: String, prefetchSize: Int) = ConfirmConsumer(this, queue, prefetchSize)
+
+suspend fun Channel.consume(queue: String, prefetchSize: Int = 0, block: suspend ConfirmConsumer.() -> Unit) =
+    this.consumer(queue, prefetchSize).use { block(it) }
