+++ src/main/kotlin/com/viartemev/thewhiterabbit/channel/Channels.kt
@@ -1,11 +1,6 @@
 package com.viartemev.thewhiterabbit.channel
 
 import com.rabbitmq.client.Channel
-import com.rabbitmq.client.Connection
-import com.viartemev.thewhiterabbit.channel.Channels.localChannels
-import com.viartemev.thewhiterabbit.channel.Channels.localConfirmChannels
-import com.viartemev.thewhiterabbit.consumer.ConfirmConsumer
-import com.viartemev.thewhiterabbit.publisher.ConfirmPublisher
 import java.util.concurrent.ConcurrentHashMap
 import kotlin.concurrent.thread
 
@@ -50,45 +45,3 @@ object Channels {
     }
 }
 
-/**
- * Create a channel with enabled publisher acknowledgements on it.Channel by channel
- * @see com.rabbitmq.client.Channel.confirmSelect()
- */
-fun Connection.createConfirmChannel(): ConfirmChannel = ConfirmChannel(this.createChannel())
-
-suspend fun Connection.confirmChannel(block: suspend ConfirmChannel.() -> Unit): ConfirmChannel {
-    var channel = Channels
-        .localConfirmChannels[Thread.currentThread()]
-    if (channel == null || !channel.isOpen) {
-        channel = Channels.UncloseableConfirmChannel(createConfirmChannel())
-        localConfirmChannels[Thread.currentThread()] = channel
-    }
-    return channel
-        .also { block(it) }
-}
-
-suspend fun Connection.channel(block: suspend Channel.() -> Unit): Channel {
-    var channel = Channels
-        .localChannels[Thread.currentThread()]
-    if (channel == null || !channel.isOpen) {
-        channel = Channels.UncloseableChannel(createChannel())
-        localChannels[Thread.currentThread()] = channel
-    }
-    return channel
-        .also { block(it) }
-}
-
-open class ConfirmChannel internal constructor(private val channel: Channel) : Channel by channel {
-    init {
-        channel.confirmSelect()
-    }
-
-    fun publisher() = ConfirmPublisher(this)
-}
-
-fun Channel.consumer(queue: String, prefetchSize: Int) = ConfirmConsumer(this, queue, prefetchSize)
-
-suspend fun ConfirmChannel.publish(block: suspend ConfirmPublisher.() -> Unit) = block(this.publisher())
-
-suspend fun Channel.consume(queue: String, prefetchSize: Int = 0, block: suspend ConfirmConsumer.() -> Unit) =
-    this.consumer(queue, prefetchSize).use { block(it) }
